avoid using the shell in this test
I can't figure out why, but in many architectures (even reproducible in my
ubuntu VirtualBox), when a shell is used to execute the command, it does not
respond to a kill signal.

A possible workaround would be to make the 'command' option a multivalued
argument, and pass the command and arguments as separate parameters to open3()
so as to avoid the shell.

e.g. see
http://www.cpantesters.org/cpan/report/00fae5e8-e0e2-11e4-ae65-7ff0e0bfc7aa